+++ EasyPost.Net/Insurance.cs
@@ -1,22 +1,36 @@
 using System.Collections.Generic;
+using Newtonsoft.Json;
 using RestSharp;
 
 namespace EasyPost
 {
     public class Insurance : Resource
     {
+        [JsonProperty("amount")]
         public string amount { get; set; }
+        [JsonProperty("from_address")]
         public Address from_address { get; set; }
+        [JsonProperty("id")]
         public string id { get; set; }
+        [JsonProperty("messages")]
         public List<string> messages { get; set; }
+        [JsonProperty("mode")]
         public string mode { get; set; }
+        [JsonProperty("provider")]
         public string provider { get; set; }
+        [JsonProperty("provider_id")]
         public string provider_id { get; set; }
+        [JsonProperty("reference")]
         public string reference { get; set; }
+        [JsonProperty("shipment_id")]
         public string shipment_id { get; set; }
+        [JsonProperty("status")]
         public string status { get; set; }
+        [JsonProperty("to_address")]
         public Address to_address { get; set; }
+        [JsonProperty("tracker")]
         public Tracker tracker { get; set; }
+        [JsonProperty("tracking_code")]
         public string tracking_code { get; set; }
 
         /// <summary>
@@ -50,7 +64,7 @@ namespace EasyPost
         ///     * {"page_size", int} Max size of list. Default to 20.
         ///     All invalid keys will be ignored.
         /// </param>
-        /// <returns>EasyPost.InsuranceCollection instance.</returns>
+        /// <returns>An EasyPost.InsuranceCollection instance.</returns>
         public static InsuranceCollection All(Dictionary<string, object> parameters = null)
         {
             parameters = parameters ?? new Dictionary<string, object>();
@@ -76,7 +90,7 @@ namespace EasyPost
         /// <returns>EasyPost.Insurance instance.</returns>
         public static Insurance Create(Dictionary<string, object> parameters)
         {
-            Request request = new Request("insurances", Method.POST);
+            Request request = new Request("insurances", Method.Post);
             request.AddBody(new Dictionary<string, object>
             {
                 {
@@ -102,7 +116,7 @@ namespace EasyPost
 
         private static Insurance SendCreate(Dictionary<string, object> parameters)
         {
-            Request request = new Request("insurances", Method.POST);
+            Request request = new Request("insurances", Method.Post);
             request.AddBody(new Dictionary<string, object>
             {
                 {