+++ EasyPost.NetFramework/AddressCollection.cs
@@ -1,10 +1,13 @@
 using System.Collections.Generic;
+using Newtonsoft.Json;
 
 namespace EasyPost
 {
     public class AddressCollection : Resource
     {
-        public List<Batch> addresses { get; set; }
+        [JsonProperty("addresses")]
+        public List<Address> addresses { get; set; }
+        [JsonProperty("has_more")]
         public bool has_more { get; set; }
     }
 }