+++ EasyPost.NetFramework/Batch.cs
@@ -1,25 +1,40 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Newtonsoft.Json;
 using RestSharp;
 
 namespace EasyPost
 {
     public class Batch : Resource
     {
+        [JsonProperty("created_at")]
         public DateTime? created_at { get; set; }
+        [JsonProperty("error")]
         public string error { get; set; }
+        [JsonProperty("id")]
         public string id { get; set; }
+        [JsonProperty("label_url")]
         public string label_url { get; set; }
+        [JsonProperty("message")]
         public string message { get; set; }
+        [JsonProperty("mode")]
         public string mode { get; set; }
+        [JsonProperty("num_shipments")]
         public int num_shipments { get; set; }
+        [JsonProperty("reference")]
         public string reference { get; set; }
+        [JsonProperty("scan_form")]
         public ScanForm scan_form { get; set; }
+        [JsonProperty("shipments")]
         public List<BatchShipment> shipments { get; set; }
+        [JsonProperty("state")]
         public string state { get; set; }
+        [JsonProperty("status")]
         public Dictionary<string, int> status { get; set; }
+        [JsonProperty("updated_at")]
         public DateTime? updated_at { get; set; }
+
 
         /// <summary>
         ///     Add shipments to this batch.
@@ -152,6 +167,31 @@ namespace EasyPost
             return request.Execute<Batch>();
         }
 
+        /// <summary>
+        ///     Create and buy a Batch in one step.
+        /// </summary>
+        /// <param name="parameters">
+        ///     Optional dictionary containing parameters to create the batch with. Valid pairs:
+        ///     * {"shipments", List&lt;Dictionary&lt;string, object&gt;&gt;} See Shipment.Create for a list of valid keys.
+        ///     * {"reference", string}
+        ///     All invalid keys will be ignored.
+        /// </param>
+        /// <returns>EasyPost.Batch instance.</returns>
+        public static Batch CreateAndBuy(Dictionary<string, object> parameters = null)
+        {
+            parameters = parameters ?? new Dictionary<string, object>();
+
+            Request request = new Request("batches/create_and_buy", Method.POST);
+            request.AddBody(new Dictionary<string, object>
+            {
+                {
+                    "batch", parameters
+                }
+            });
+
+            return request.Execute<Batch>();
+        }
+
 
         /// <summary>
         ///     Retrieve a Batch from its id.
@@ -180,7 +220,7 @@ namespace EasyPost
         ///     * {"page_size", int} Max size of list. Default to 20.
         ///     All invalid keys will be ignored.
         /// </param>
-        /// <returns>EasyPost.BatchCollection instance.</returns>
+        /// <returns>An EasyPost.BatchCollection instance.</returns>
         public static BatchCollection All(Dictionary<string, object> parameters = null)
         {
             parameters = parameters ?? new Dictionary<string, object>();