+++ EasyPost.NetFramework/CarrierAccount.cs
@@ -1,30 +1,41 @@
 ﻿using System;
 using System.Collections.Generic;
+using Newtonsoft.Json;
 using RestSharp;
 
 namespace EasyPost
 {
     public class CarrierAccount : Resource
     {
+        [JsonProperty("created_at")]
         public DateTime? created_at { get; set; }
+        [JsonProperty("credentials")]
         public Dictionary<string, object> credentials { get; set; }
+        [JsonProperty("description")]
         public string description { get; set; }
+        [JsonProperty("id")]
         public string id { get; set; }
+        [JsonProperty("readable")]
         public string readable { get; set; }
+        [JsonProperty("reference")]
         public string reference { get; set; }
+        [JsonProperty("test_credentials")]
         public Dictionary<string, object> test_credentials { get; set; }
+        [JsonProperty("type")]
         public string type { get; set; }
+        [JsonProperty("updated_at")]
         public DateTime? updated_at { get; set; }
 
         /// <summary>
         ///     Remove this CarrierAccount from your account.
         /// </summary>
-        public void Destroy()
+        /// <returns>Whether the request was successful or not.</returns>
+        public bool Delete()
         {
             Request request = new Request("carrier_accounts/{id}", Method.DELETE);
             request.AddUrlSegment("id", id);
 
-            request.Execute();
+            return request.Execute();
         }
 
         /// <summary>
@@ -76,7 +87,7 @@ namespace EasyPost
         ///     List all available carrier accounts.
         /// </summary>
         /// <returns>A list of EasyPost.CarrierAccount instances.</returns>
-        public static List<CarrierAccount> List()
+        public static List<CarrierAccount> All()
         {
             Request request = new Request("carrier_accounts");
             return request.Execute<List<CarrierAccount>>();