+++ EasyPost.Tests.NetFramework/TrackerTest.cs
@@ -1,70 +1,97 @@
 using System.Collections.Generic;
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace EasyPost.Tests.Net
+namespace EasyPost.Tests.NetFramework
 {
     [TestClass]
     public class TrackerTest
     {
         [TestInitialize]
-        public void Initialize() => ClientManager.SetCurrent("NvBX2hFF44SVvTPtYjF0zQ");
-
-        [Ignore]
-        [TestMethod]
-        public void TestCreateAndRetrieve()
+        public void Initialize()
         {
-            const string carrier = "USPS";
-            const string trackingCode = "EZ1000000001";
+            TestSuite.SetUp(TestSuiteApiKey.Test);
+        }
 
-            Tracker tracker = Tracker.Create(carrier, trackingCode);
-            Assert.AreEqual(tracker.tracking_code, trackingCode);
-            Assert.IsNotNull(tracker.est_delivery_date);
-            Assert.IsNotNull(tracker.carrier);
-            Assert.IsNotNull(tracker.public_url);
-
-            Assert.AreEqual(Tracker.Retrieve(tracker.id).id, tracker.id);
+        private static Tracker CreateBasicTracker()
+        {
+            return Tracker.Create(Fixture.Usps, "EZ1000000001");
         }
 
         [TestMethod]
-        public void TestCreateTrackerList()
+        public void TestCreate()
         {
-            string[] trackingCodes =
-            {
-                "EZ1000000001", "EZ1000000002", "EZ1000000003"
-            };
-            Dictionary<string, object> trackers = new Dictionary<string, object>();
+            Tracker tracker = CreateBasicTracker();
 
-            for (int i = 0; i < trackingCodes.Length; i++)
-            {
-                trackers.Add(i.ToString(), new Dictionary<string, object>
-                {
-                    {
-                        "tracking_code", trackingCodes[i]
-                    },
-                    {
-                        "carrier", "USPS"
-                    }
-                });
-            }
-
-            bool response = Tracker.CreateList(trackers);
-            Assert.IsTrue(response);
+            Assert.IsInstanceOfType(tracker, typeof(Tracker));
+            Assert.IsTrue(tracker.id.StartsWith("trk_"));
+            Assert.AreEqual("pre_transit", tracker.status);
         }
 
         [TestMethod]
-        public void TestList()
+        public void TestRetrieve()
         {
-            TrackerList trackerList = Tracker.List(new Dictionary<string, object>
+            // Test trackers cycle through their "dummy" statuses automatically, the created and retrieved objects may differ
+            Tracker tracker = CreateBasicTracker();
+
+            Tracker retrievedTracker = Tracker.Retrieve(tracker.id);
+
+            Assert.IsInstanceOfType(retrievedTracker, typeof(Tracker));
+            Assert.AreEqual(tracker.id, retrievedTracker.id);
+        }
+
+        [TestMethod]
+        public void TestAll()
+        {
+            TrackerCollection trackerCollection = Tracker.All(new Dictionary<string, object>
             {
                 {
-                    "page_size", 1
+                    "page_size", Fixture.PageSize
                 }
             });
-            Assert.AreNotEqual(0, trackerList.trackers.Count);
 
-            TrackerList nextTrackerList = trackerList.Next();
-            Assert.AreNotEqual(trackerList.trackers[0].id, nextTrackerList.trackers[0].id);
+            List<Tracker> trackers = trackerCollection.trackers;
+
+            Assert.IsTrue(trackers.Count <= Fixture.PageSize);
+            Assert.IsNotNull(trackerCollection.has_more);
+            foreach (var tracker in trackers)
+            {
+                Assert.IsInstanceOfType(tracker, typeof(Tracker));
+            }
+        }
+
+        [TestMethod]
+        public void TestCreateList()
+        {
+            bool success = Tracker.CreateList(new Dictionary<string, object>
+            {
+                {
+                    "0", new Dictionary<string, object>
+                    {
+                        {
+                            "tracking_code", "EZ1000000001"
+                        }
+                    }
+                },
+                {
+                    "1", new Dictionary<string, object>
+                    {
+                        {
+                            "tracking_code", "EZ1000000002"
+                        }
+                    }
+                },
+                {
+                    "2", new Dictionary<string, object>
+                    {
+                        {
+                            "tracking_code", "EZ1000000003"
+                        }
+                    }
+                },
+            });
+
+            // This endpoint returns nothing so we assert the function returns true
+            Assert.IsTrue(success);
         }
     }
 }